fix stack overflow in dexie TODO: more research here

+++ src/node.rs
@@ -21,8 +21,8 @@ impl SourceLocation {
 
 #[derive(PartialEq, Debug, Clone, Copy)]
 pub struct Position {
-    pub line: usize,
-    pub column: usize,
+    pub line: u32,
+    pub column: u32,
 }
 
 impl ::std::fmt::Display for Position {
@@ -35,7 +35,7 @@ impl Position {
     pub fn start() -> Self {
         Self { line: 0, column: 0 }
     }
-    pub fn new(line: usize, column: usize) -> Self {
+    pub fn new(line: u32, column: u32) -> Self {
         Self { line, column }
     }
 }
